[refactor] validator 테스트 코드 test_validator로 분리
