+++ users/tests/test_user_sign_up.py
@@ -1,10 +1,5 @@
-from rest_framework.exceptions import ValidationError
-from django.test import TestCase
 from rest_framework import status
 from rest_framework.test import APITestCase
-
-from users.models import User
-from users.validator import not_exist_user_nickname, not_exist_user_email
 
 test_user_data = {
     "username": "test1",
@@ -23,16 +18,3 @@ class TestUserSignUp(APITestCase):
         self.assertEqual(response.data['isSuccess'], True)
         self.assertEqual(response.data['message'], 'OK')
         self.assertTrue('id' in response.data['result'])
-
-
-class TestValidator(TestCase):
-    def setUp(self):
-        User.objects.create(**test_user_data)
-
-    def test_not_exist_nickname(self):
-        self.assertIsNone(not_exist_user_nickname(nickname='unique_nickname'))
-        self.assertRaises(ValidationError, not_exist_user_nickname, nickname=test_user_data['nickname'])
-
-    def test_not_exist_user_email(self):
-        self.assertIsNone(not_exist_user_email(email='unique_email'))
-        self.assertRaises(ValidationError, not_exist_user_email, email=test_user_data['email'])
